requirements/testing.txt
# Testing
coverage==3.7.1
coveralls==0.5
Flask-Testing==0.4.2
mock==1.1.2
pytest==2.7.2
vcrpy==1.6.0
codecov
